requirements.txt
Django 4.2.8
numpy 1.26.3
pandas 2.1.4
Pillow 10.1.0
pip 21.1.1
psycopg2 2.9.9
qrcode 7.4.2
requests 2.31.0
sqlparse 0.4.4
XlsxWriter 3.1.9
html2pdf
